fixed monstervalley html structure. Grass div is now the valley.
