+++ src/components/animations/monster/MonsterValley.module.css
@@ -1,12 +1,28 @@
 .valley {
+  height: 300px;
+  background-image: url('../../../assets/footer/grass.png');
+  background-repeat: repeat-x;
+  background-size: auto 100%;
   display: flex;
   justify-content: space-evenly;
-  margin-bottom: -450px;
-  height: 20rem;
-  position: relative;
-  z-index: 0;
+  align-items: center;
+}
+
+.valley button {
+  margin-top: 14rem;
+  position: absolute;
+  z-index: 3;
+  background-color: transparent;
+  border: none;
+  text-shadow: 1px 1px 2px #ffffff;
+  font-size: larger;
+  font-weight: 900;
+  cursor: pointer;
 }
 
 .valley img {
   height: 200px;
+  margin-top: 13rem;
+  position: relative;
+  z-index: 0;
 }
