+++ src/components/footer/LandingFooter.module.css
@@ -4,27 +4,3 @@
   overflow: hidden;
   padding-top: 2rem;
 }
-
-.grass {
-  height: 300px;
-  background-image: url('../../assets/footer/grass.png');
-  background-repeat: repeat-x;
-  background-size: auto 100%;
-
-  /* positions button only */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.grass button {
-  margin-top: 14rem;
-  position: relative;
-  z-index: 3;
-  background-color: transparent;
-  border: none;
-  text-shadow: 1px 1px 2px #ffffff;
-  font-size: larger;
-  font-weight: 900;
-  cursor: pointer;
-}
